refactor: split CpSatSolver into a package
Decompose the 1879-line cp_sat_solver.py into a focused package with
five submodules:

  cp_sat_solver/
    __init__.py        # re-exports CpSatSolver
    _types.py          # _TaskVariables, _AlternativeVariables,
                       # _safe_name, _load_cp_model (54 lines)
    _model_helpers.py  # _ModelHelpersMixin: intervals, windows,
                       # cumulatives, dependency edges (626 lines)
    _heuristic.py      # _HeuristicMixin: candidate orderings, warm
                       # start, AddHint plumbing (196 lines)
    _objective.py      # _ObjectiveMixin: job-completion + objective
                       # IntVars and weighted expression (222 lines)
    solver.py          # CpSatSolver class: __init__, _allocate_tasks,
                       # _configure_solver (969 lines)

CpSatSolver inherits BaseSolver plus the three private mixins, so the
class surface stays unchanged and every `self.method(...)` call
resolves the same as before. The package import path
(`from frost_planner.solver.cp_sat_solver import CpSatSolver`) is
preserved through __init__.py, so existing call sites in factory.py
and the test suite don't need to change.

All 140 tests pass.
